Extracted create_build_dir
